Move `MomentState` & friends to the new shared sub-package
