+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/EventStream.kt
@@ -1,5 +1,7 @@
 package io.github.oxidefrp.core
 
+import io.github.oxidefrp.core.shared.MomentState
+import io.github.oxidefrp.core.shared.State
 import io.github.oxidefrp.core.shared.StateScheduler
 import io.github.oxidefrp.core.shared.StateSchedulerLayer
