+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/Io.kt
@@ -1,5 +1,7 @@
 package io.github.oxidefrp.core
 
+import io.github.oxidefrp.core.shared.State
+
 data class RealWorld(
     val files: Map<String, String>,
 ) {
