+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/MomentIo.kt
@@ -1,3 +1,5 @@
 package io.github.oxidefrp.core
 
+import io.github.oxidefrp.core.shared.MomentState
+
 typealias MomentIo<A> = MomentState<RealWorld, A>
