+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/shared/MomentState.kt
@@ -1,6 +1,12 @@
-package io.github.oxidefrp.core
+package io.github.oxidefrp.core.shared
 
+import io.github.oxidefrp.core.Moment
+import io.github.oxidefrp.core.Signal
+import io.github.oxidefrp.core.Time
+import io.github.oxidefrp.core.pullOf
+import io.github.oxidefrp.core.shared.State
 import io.github.oxidefrp.core.shared.StateSchedulerLayer
+import io.github.oxidefrp.core.shared.StateStructure
 
 abstract class MomentState<S, out A> {
     companion object {
