+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/shared/StateScheduler.kt
@@ -3,10 +3,7 @@ package io.github.oxidefrp.core.shared
 import io.github.oxidefrp.core.Cell
 import io.github.oxidefrp.core.EventStream
 import io.github.oxidefrp.core.Moment
-import io.github.oxidefrp.core.MomentState
 import io.github.oxidefrp.core.Signal
-import io.github.oxidefrp.core.State
-import io.github.oxidefrp.core.StateStructure
 import io.github.oxidefrp.core.mapNotNull
 import io.github.oxidefrp.core.orElse
 import io.github.oxidefrp.core.pullOf
